% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseComputations.R
\docType{data}
\name{baseComputations}
\alias{baseComputations}
\title{baseComputations}
\format{
An object of class \code{list} of length 7.
}
\usage{
baseComputations
}
\description{
A list of computation objects to be used by \code{\link{run}} via the `whatToRun` parameter.
}
\details{
Each computation object is a list with a setup and a compute function. These are called at the appropriate times
during a piktests \code{\link{run}}.
}
\author{
Pascal Führlich
}
\keyword{datasets}
