% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLogs.R
\name{compareLogs}
\alias{compareLogs}
\title{compareLogs}
\usage{
compareLogs(pathToLogA, pathToLogB)
}
\arguments{
\item{pathToLogA}{Path to the first log file, shown on the left hand side in the resulting diff.}

\item{pathToLogB}{Path to the second log file, shown on the right hand side in the resulting diff.}
}
\description{
Show side-by-side diff of the two given log files.
}
\details{
Runtimes are removed from logs before comparison.
}
