% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRunFolder.R
\name{createRunFolder}
\alias{createRunFolder}
\title{createRunFolder}
\usage{
createRunFolder(
  computationNames = c("magpiePrep", "remindPrep"),
  piktestsFolder = getwd(),
  runFolder = NULL
)
}
\arguments{
\item{computationNames}{These are appended to the created folder's name.}

\item{piktestsFolder}{A new folder for this piktests run is created in the given directory.}

\item{runFolder}{Path where a folder for this piktests run should be created. Generally should be left as default,
which creates a folder name based on the current date and computationNames.}
}
\description{
Constructs a folder name based on current data and existing folders
}
\author{
Jan Philipp Dietrich
}
