% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piktests-package.R
\docType{package}
\name{piktests-package}
\alias{piktests}
\alias{piktests-package}
\title{piktests: Run PIK Integration Tests}
\description{
This package includes integration tests for selected models and packages related to those models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piktests}
}

}
\author{
\strong{Maintainer}: Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
}

}
