% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(path = ".")
}
\arguments{
\item{path}{Path to a folder. Here, a new folder named after the current date/time will be created for this test run.}
}
\description{
Runs all integration tests available in a mostly isolated runtime environment.
}
\details{
A new folder will be created for each run, named after the current date/time. In this folder a madratCacheFolder and
a madratOutputFolder are created and used while running the tests. A runtimeWorkingDirectory is also created and
while running the tests this will be the working directory. The original madrat configuration and working directory
will be restored afterwards.
}
