% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInNewRSession.R
\name{runInNewRSession}
\alias{runInNewRSession}
\title{withr local_tempfile}
\usage{
runInNewRSession(
  workFunction,
  arguments = list(),
  workFileName = tempfile("workFile-", getwd(), ".rds"),
  ...,
  cleanupWorkFile = TRUE,
  useSbatch = FALSE,
  sbatchArguments = c(paste0("--job-name=", shQuote(paste0("runInNewRSession-",
    basename(workFileName)))), paste0("--output=", shQuote(paste0("runInNewRSession-",
    basename(workFileName), ".log"))), "--mail-type=END", "--qos=priority",
    "--mem=32000")
)
}
\description{
withr local_tempfile
}
