% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInRenv.R
\name{runInRenv}
\alias{runInRenv}
\title{runInRenv}
\usage{
runInRenv(useSlurm = TRUE)
}
\arguments{
\item{useSlurm}{Whether to start the tests via sbatch (default, run in background) or directly in the current shell.}
}
\description{
Creates a test run folder, sets up a fresh renv and installs needed packages in it. Then calls piktests:::run() to
run the actual tests.
}
