% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWithComparison.R
\name{runWithComparison}
\alias{runWithComparison}
\title{runWithComparison}
\usage{
runWithComparison(
  renvInstallPackages,
  piktestsFolder = getwd(),
  whatToRun = c("remind-preprocessing", "magpie-preprocessing")
)
}
\arguments{
\item{renvInstallPackages}{Only in the second run, after installing
other packages, `renv::install(renvInstallPackages)` is called.}

\item{piktestsFolder}{A new folder is created in the given directory. In that folder two folders called "old"
and "new" are created which contain the actual piktests runs.}

\item{whatToRun}{A character vector defining what tests to run. See default value for a list of all possible tests.}
}
\value{
Invisibly, the path to the folder holding the two actual piktests runs.
}
\description{
Starts two piktests runs, one with default packages and another one with renvInstallPackages installed, so they can
be compared. Run this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").
Use `piktests::comparePreprocessingLogs` for comparison after all runs are completed.
}
\seealso{
\code{\link{comparePreprocessingLogs}}, \code{\link{run}}
}
\author{
Pascal Führlich
}
