% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRenv.R
\name{setupRenv}
\alias{setupRenv}
\title{setupRenv}
\usage{
setupRenv(targetFolder, gitCloneRepos = NULL, renvInstallPackages = NULL)
}
\arguments{
\item{targetFolder}{Where to setup the renv.}

\item{gitCloneRepos}{A named list where the value is a git url and the corresponding name is a path where to clone
that repo.}

\item{renvInstallPackages}{After installing other packages, renv::install(renvInstallPackages) is called.
Use this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}
}
\description{
Sets up a fresh renv and installs the required packages in it.
}
\author{
Pascal Führlich
}
