% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePreprocessingLogs.R
\name{comparePreprocessingLogs}
\alias{comparePreprocessingLogs}
\title{comparePreprocessingLogs}
\usage{
comparePreprocessingLogs(preprocessingName, runFolder = getwd())
}
\arguments{
\item{preprocessingName}{Name of the preprocessing you want to compare logs for, e.g. "magpie" or "remind".}

\item{runFolder}{The path to a run folder created by `piktests::runWithComparison`.}
}
\description{
Convenience function for comparing logs after running `piktests::runWithComparison`.
}
\seealso{
\code{\link{runWithComparison}}
}
\author{
Pascal Führlich
}
