% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  renvInstallPackages = NULL,
  computations = c("magpiePrep", "remindPrep"),
  piktestsFolder = getwd(),
  runFolder = NULL,
  jobNameSuffix = "",
  executionMode = c("slurm", "directly"),
  localCache = TRUE
)
}
\arguments{
\item{renvInstallPackages}{After installing other packages, renv::install(renvInstallPackages) is called.
Use this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}

\item{computations}{A named list of "computations", or names of computations predefined in
\code{\link{baseComputations}}. A computation consists of a setup and a compute function.
See example for a valid computation list.}

\item{piktestsFolder}{A new folder for this piktests run is created in the given directory.}

\item{runFolder}{Path where a folder for this piktests run should be created. Generally should be left as default,
which creates a folder name based on the current date, time, and the computation names.}

\item{jobNameSuffix}{A suffix to be appended to the SLURM job's name.}

\item{executionMode}{Determines how long running jobs are started. One of "slurm", "directly"}

\item{localCache}{If TRUE (default) use a new and empty cache folder, otherwise `getConfig("cachefolder")`.}
}
\value{
Invisibly, the path to the folder holding everything related to this piktests run.
}
\description{
Runs integration tests in an isolated runtime environment.
}
\details{
The preconfigured madrat source and mapping folders are used, otherwise a subfolder of the newly created run folder
is used as madrat mainfolder.
}
\examples{
\dontrun{
piktests::run(renvInstallPackages = c("tscheypidi/madrat", "magclass@6.0.9"),
              computations = list(testComputation = list(setup = function() message("Hello"),
                                                         compute = function() message("World"))),
               executionMode = "directly",
               localCache = FALSE)
}

}
\seealso{
\code{\link{baseComputations}}
}
\author{
Pascal Führlich, Jan Philipp Dietrich
}
