% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  renvInstallPackages = NULL,
  piktestsFolder = getwd(),
  whatToRun = c("remind-preprocessing", "magpie-preprocessing"),
  runFolder = file.path(piktestsFolder, format(Sys.time(), "\%Y_\%m_\%d-\%H_\%M")),
  runInNewRSession = callr::r
)
}
\arguments{
\item{renvInstallPackages}{After installing other packages, renv::install(renvInstallPackages) is called.
Use this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}

\item{piktestsFolder}{A new folder for this piktests run is created in the given directory.}

\item{whatToRun}{A character vector defining what tests to run. See default value for a list of all possible tests.}

\item{runFolder}{In general this should be left as default. Where the folder for this piktests run should be created.}

\item{runInNewRSession}{Exists for testing. A function like `callr::r` taking a function and arguments to execute
in a new R session.}
}
\value{
Invisibly, the path to the folder holding everything related to this piktests run.
}
\description{
Runs integration tests in an isolated runtime environment.
}
\details{
A madratCacheFolder and a madratOutputFolder are created and used while running the tests. The non-public magpie
preprocessing repo `git@gitlab.pik-potsdam.de:landuse/preprocessing-magpie.git` is cloned, so you need access to it.
}
\author{
Pascal Führlich
}
