% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(useSbatch, madratConfig, runFolder)
}
\arguments{
\item{useSbatch}{Whether to start the tests via sbatch (run in background) or directly in the current shell.}

\item{madratConfig}{The madrat configuration to use.}

\item{runFolder}{Path to the folder where everything related to this piktests run is stored. Must be an renv project.}
}
\description{
This function is only meant to be run by piktests::runInRenv. Runs all integration tests in a mostly isolated runtime
environment.
}
\details{
A runtimeWorkingDirectory is created and used as the working directory while running the tests, also a
madratCacheFolder and a madratOutputFolder are created and used while running the tests.
}
