% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  renvInstallPackages = NULL,
  piktestsFolder = getwd(),
  whatToRun = c("remind-preprocessing", "magpie-preprocessing")
)
}
\arguments{
\item{renvInstallPackages}{After installing other packages, renv::install(renvInstallPackages) is called.
Use this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}

\item{piktestsFolder}{A new folder for this piktests run is created in the given directory.}

\item{whatToRun}{A character vector defining what tests to run. See default value for a list of all possible tests.}
}
\description{
Runs integration tests in an isolated runtime environment.
}
\details{
A madratCacheFolder and a madratOutputFolder are created and used while running the tests.
}
