% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  renvInstallPackages = NULL,
  computations = baseComputations[c("magpiePrep", "remindPrep")],
  piktestsFolder = getwd(),
  runFolder = NULL,
  jobNameSuffix = "",
  executionMode = c("slurm", "directly"),
  localCache = TRUE
)
}
\arguments{
\item{renvInstallPackages}{After installing other packages, renv::install(renvInstallPackages) is called.
Use this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}

\item{piktestsFolder}{A new folder for this piktests run is created in the given directory.}

\item{runFolder}{Path where a folder for this piktests run should be created. Generally should be left as default,
which creates a folder name based on the current date, time, and computationNames.}

\item{jobNameSuffix}{A suffix to be appended to the SLURM job's name.}

\item{executionMode}{Determines how long running jobs are started. One of "slurm", "directly"}

\item{localCache}{If TRUE (default) use a new and empty cache folder, otherwise `getConfig("cachefolder")`.}

\item{computationNames}{A subset of names(piktests::computations). The setup and compute functions of these
computations are executed.}
}
\value{
Invisibly, the path to the folder holding everything related to this piktests run.
}
\description{
Runs integration tests in an isolated runtime environment.
}
\details{
A madratCacheFolder and a madratOutputFolder are created and used while running the tests. The non-public magpie
preprocessing repo `git@gitlab.pik-potsdam.de:landuse/preprocessing-magpie.git` is cloned, so you need access to it.
}
\seealso{
\code{\link{computations}}
}
\author{
Pascal Führlich, Jan Philipp Dietrich
}
