% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInNewRSession.R
\name{runInNewRSession}
\alias{runInNewRSession}
\title{runInNewRSession}
\usage{
runInNewRSession(
  workFunction,
  arguments = list(),
  workFileName = tempfile("workFile-", getwd(), ".rds"),
  ...,
  cleanupWorkFile = TRUE,
  useSbatch = FALSE,
  sbatchArguments = c(paste0("--job-name=", shQuote(paste0("runInNewRSession-",
    basename(workFileName)))), paste0("--output=", shQuote(paste0("runInNewRSession-",
    basename(workFileName), ".log"))), "--mail-type=END", "--qos=priority",
    "--mem=32000")
)
}
\arguments{
\item{workFunction}{The function to execute in the new R session. TODO}

\item{arguments}{A list of arguments that is passed to workFunction.}

\item{workFileName}{The file name of an RDS file containing workFunction and arguments. This file is read in the new
R session.}

\item{...}{Additional arguments passed to system2. The most useful arguments are probably stdout and stderr, see
documentation for system2.}

\item{cleanupWorkFile}{Whether to delete the workFile after it has been read in the new R session.}

\item{useSbatch}{Whether to run the new R session in the background using sbatch.}

\item{sbatchArguments}{Arguments passed to sbatch. A --wrap argument must not be passed. A --wrap argument running
the given function in a new R session is automatically appended.}
}
\value{
The result of the system2 call which is calling Rscript directly or via sbatch.
}
\description{
Runs the given function with the given arguments in a new R session.
}
\details{
An RDS file containing the function and args is created, then a new R session is started via system2. In this new
session the RDS file is read and the function executed with the arguments.
}
