% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInRenv.R
\name{runInRenv}
\alias{runInRenv}
\title{runInRenv}
\usage{
runInRenv(
  work,
  ...,
  targetFolder = ".",
  renvLockfile = NULL,
  renvPreInstall = NULL
)
}
\arguments{
\item{work}{A function that should be executed in the newly created renv.}

\item{...}{Arguments passed to \code{\link{runLongJob}}. `renvToLoad` will be set automatically.}

\item{targetFolder}{Where the renv is created.}

\item{renvLockfile}{Optional. An renv lockfile used to create the renv. If NULL auto-detect dependencies instead.}

\item{renvPreInstall}{Optional. A function that is run right after initializing the renv, can be used to apply
system specific fixes.}
}
\description{
Experimental. Create an renv with all packages needed to run the given function and run that function in that renv.
}
