% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLongJob.R
\name{runLongJob}
\alias{runLongJob}
\title{runLongJob}
\usage{
runLongJob(
  workFunction,
  arguments = NULL,
  workingDirectory = getwd(),
  renvToActivate = NULL,
  madratConfig = NULL,
  jobName = opts_slurmR$get_job_name(),
  mode = c("sbatch", "parallel", "sequential")
)
}
\description{
Run a function in a new R session.
}
