% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLongJob.R
\name{runLongJob}
\alias{runLongJob}
\title{runLongJob}
\usage{
runLongJob(
  workFunction,
  arguments = list(),
  workingDirectory = getwd(),
  renvToLoad = NULL,
  madratConfig = NULL,
  jobName = opts_slurmR$get_job_name(),
  executionMode = c("slurm", "background", "directly")
)
}
\arguments{
\item{workFunction}{This function will be run in a new R session, so it must use `::` whenever package functions are
used. Also it cannot refer to variables in the outer scope, use the next parameter (arguments) to pass them.}

\item{arguments}{A list of arguments passed to workFunction.}

\item{workingDirectory}{The working directory in which workFunction will be called.}

\item{renvToLoad}{The renv project to load before running workFunction.}

\item{madratConfig}{A madrat config (as returned by `madrat::getConfig()`) to be used when running workFunction.}

\item{jobName}{The output file will be called `<jobName>.log`. If executionMode is "slurm" this is also the name
of the slurm job.}

\item{executionMode}{Determines how workFunction is started.
"slurm" -> `slurmR::Slurm_lapply`, "background" -> `callr::r_bg`, "directly" -> `callr::r`}
}
\description{
Run a function in a new R session, per default via SLURM (see executionMode).
}
\author{
Pascal Führlich
}
