% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRenv.R
\name{setupRenv}
\alias{setupRenv}
\title{setupRenv}
\usage{
setupRenv(runFolder, renvInstallPackages = NULL)
}
\arguments{
\item{runFolder}{Where to setup the renv.}

\item{renvInstallPackages}{After installing other packages, renv::install(renvInstallPackages) is called.
Use this to test changes in your fork by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}
}
\description{
Sets up a fresh renv and installs the required packages in it.
}
