% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRenv.R
\name{setupRenv}
\alias{setupRenv}
\title{setupRenv}
\usage{
setupRenv(
  targetFolder,
  computationNames = c("madratExample", "magpiePreprocessing"),
  renvInstallPackages = NULL
)
}
\arguments{
\item{targetFolder}{Where to setup the renv.}

\item{computationNames}{A subset of names(piktests::computations). The setup functions of these computations are run.}

\item{renvInstallPackages}{renv::install(renvInstallPackages) is called. Use this to test changes in your fork
by passing "<gituser>/<repo>" (e.g. "pfuehrlich-pik/madrat").}
}
\description{
Sets up a fresh renv and installs the required packages in it.
}
\details{
This function should be called in a fresh R session (e.g. via callr::r), because setting up an renv involves
changing critical aspects of your R session like your libpaths.
}
\examples{
\dontrun{
callr::r(piktests:::setupRenv,
         list(targetFolder = tempdir(),
              computationNames = c("madratExample", "magpiePreprocessing"),
              renvInstallPackages = c("tscheypidi/madrat", "magclass@6.0.9")),
         show = TRUE)
}
}
\seealso{
\code{\link{computations}}
}
\author{
Pascal Führlich
}
